bump rust version from 1.63 to 1.66 and misc project cleanups (#1459)

+++ utils/releaser/src/android/mod.rs
@@ -50,7 +50,7 @@ fn release_gh(gh: &Github) {
     client
         .upload_release_asset(
             &lb_repo(),
-            release.id as u64,
+            release.id,
             "lockbook-android.apk",
             "application/vnd.android.package-archive",
             file,
