+++ utils/releaser/src/linux/cli.rs
@@ -70,7 +70,7 @@ pub fn upload(gh: &Github) {
     client
         .upload_release_asset(
             &lb_repo(),
-            release.id as u64,
+            release.id,
             "lockbook-cli",
             "application/octet-stream",
             file,
